backend/requirements.txt
asgiref==3.8.1
Django==4.2
django-cors-headers==4.5.0
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
gunicorn==23.0.0
importlib_metadata==8.5.0
Markdown==3.7
packaging==24.2
promise==2.3
psycopg2-binary==2.9.10
Pygments==2.18.0
PyJWT==1.7.1
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
typing_extensions==4.12.2
zipp==3.21.0
